Optimize exponent() for even numbers
This commit optimizes exponent() for even numbers by using the
exponentiation by squaring method detailed here:
https://en.wikipedia.org/wiki/Exponentiation_by_squaring

The tests do not indicate that it is optimized yet, so there
appears to be room for further optimization. However, the changes
have reduced the number of calls to the call stack by dividing
the exponent on each call by two as the decrementer rather than
subtracting one as the decrementer.